fix pr 8270

+++ src/Orchard.Web/Modules/Orchard.Autoroute/Services/AutorouteService.cs
@@ -180,7 +180,8 @@ namespace Orchard.Autoroute.Services {
         }
 
         public void RemoveAliases(AutoroutePart part) {
-            _aliasService.Delete(part.Path, AliasSource);
+            if (part != null)
+                _aliasService.Delete(part.Path, AliasSource);
         }
 
         public string GenerateUniqueSlug(AutoroutePart part, IEnumerable<string> existingPaths) {
